AuthenticatedControllerAction sets Raven current_user for APIs
